[arcane,accelerator] Ajoute fonctions pour récupérer la 'stream' native associé à une 'RunQueue'.
